Flask app structure refactor
- Create a `flask_app` object in __init__.py per recommendation of
  http://flask.pocoo.org/docs/0.12/patterns/packages/

- Update the uwsgi configure to point to the new app object.

- Deprecate the sqre-uservice-ccutter command; instead we'll use the
  `flask` app runner for development. This is documented in the README.

- Refactor flask app creation into createapp module.

- Refactored cache update into templatecache module.

- Use a blueprint in uservice_ccutter/routes/ to set up the APIs.
  All the routes are there.

- This lets us deprecate the server.py module
