+++ uservice_ccutter/__init__.py
@@ -1,5 +1,8 @@
 """SQuaRE cookiecutter service (api.lsst.codes-compliant).
 """
-from .server import server, standalone
-from .projecturls import PROJECTURLS
-__all__ = ["server", "standalone", "PROJECTURLS"]
+
+__all__ = ['flask_app']
+
+from .createapp import create_flask_app
+
+flask_app = create_flask_app()
