+++ uservice_ccutter/routes/root.py
@@ -1,0 +1,10 @@
+__all__ = ['healthcheck']
+
+from . import api
+
+
+@api.route("/")
+def healthcheck():
+    """Root endpoint used for Kubernetes  health checks.
+    """
+    return "OK"
