+++ uwsgi.ini
@@ -5,7 +5,6 @@ virtualenv = %(_)
 endif =
 http = :5000
 module = uservice_ccutter
+callable = flask_app
 ; *Really* increase the timeout
 harakiri = 600
-; You need the parens, because you want server's return value (a Flask app)
-callable = server()
